requirements.txt
openpyxl
requests
python-dotenv
playwright
